Fixed problem with crashing on 64-bit Linux & possibly others.
Quite frankly we don't know how it was running anywhere else, as Qt made
a subtle change in how their QApplication objects were created without
making it known far and wide. Maybe they did, and I just missed it. But
the lesson learned is that for want of an ampersand, a kingdom was lost.