More gamepad work.
Gamepad configuration and usability is now working, though there are
some limitations: you must be careful to configure the pads you want to
be pad 1 & 2. You can mix keyboard keys, but not have multiply defined
keys (yet). There is no joypad specific configuration, though that may
come in the future if there is demand for it.