Fixed fullscreen mode, shrunk screen width.
Full screeen mode now correctly centers the image on wide screen
monitors. In addition, the virtual screen width has been reduced in size
back to what you would see on a TV with a modest amount of overscan.
Assuming all is well with this commit, this will be our 2.1.0 release.
Yes, there are still some regressions in a few games that we are aware
of, but the improvements over the last official release are too great to
hold this back any longer. :-)