Various changes to improve code readability, added mouse hiding.
First steps towards using the variables exposed in memory.cpp; there's
still a long way to go. Reverted the instruction timings back to the old
and incorrect shorter durations; this fixes sound in Rayman and FACTS.
Also added mouse hiding: if the mouse is sitting on top of the VJ window
and it hasn't moved in around a second, it is hidden until you move it
again. This behaviour only happens when the emulated Jaguar is running;
at all other times it is quiescent.