Fixes for fullscreen mode, initial stab at multiple controller configs
Who knew that you could attach QActions to the main window? Of course,
you'd have to be a complete idiot not to see something so completely and
blatantly obvoius! :-P The preceeding was a cheap dig at the kind folks
who hang out on #qt on Freenode, who grudgingly provided that little
tidbit of information as if it was the most obvious thing in the world.

The upshot is that the toolbar is gone from top of the screen in full
screen mode now, and that should make lots of people happy. :-) Also,
the initial GUI implementation for multiple controller profiles is in.
More to follow.