Fixed updated joystick handling, first major stab at gamepad profiles.
There was a problem in JERRY's read byte routine, it was getting bad
data as a result of the conversion of the joystick read routines to
word data size only. Apparently some programs (like Rayman) read
JOYSTICK a byte at a time. :-P

Also, this marks the first stable commit of the the new gamepad profile
system. There's still work to do, but this should make it much nicer
for people who use VJ with multiple input devices.