Added user controlled blitter logging.
Press F10 to have the blitter do a "short" dump of the blitter command;
this also dumps various blitter registers as well. Press F12 to have VJ
log the internal states of the blitter as well. This is pretty much only
useful if you're trying to fix problems with the blitter code. ;-) Also,
you *must* have logging turned on with the --log option to get any input.
Those keys only turn on logging. There's no way to turn it off once it's
started, other than exiting the emulator.