Fixed controller profile system.
What this means is that you have to set up profiles for controllers you
plug in now. Virtual Jaguar will try to do the right thing, but it's
still possible to make it screw up and do the wrong thing. Hopefully
this will be fixed in a future update. :-P