Made keyboard handling a bit more sane.
What this means is that now, if there are no gamepads plugged in, it
will add the keyboard as a device plugged into slot #0 (CONTROLLER1).
Also, if you have any devices plugged in it will also honor whatever
choice you have made WRT the keyboard and attempt to connect it to
whichever preferred slot you have set it to.

Assuming testing goes OK, this will be the 2.1.1 release. :-D